[CP] Import a CSS from a dependency
In this commit we add a Clay Chart widget to the portlet. The Clay Chart
package comes with a bundled CSS file that enhances the display of the
chart widgets, so we use import to include it in our page.

You can deploy this commit and should see a chart with thin axis lines.
If you see thick axis lines something went wrong.