[CP] Import a JSON file
Use the JSON loader to do an import of a .json file and be able to
handle its contents as a JavaScript object, much in the same way as
Node.js lets you require .json files.

You can deploy this commit and you should see the contents of file.json
right below the button.